@@ -11,24 +11,22 @@
     }
   ],
   "require": {
-    "php": "^5.6||^7.0",
-    "symfony/dependency-injection": "^2.7|^3.0|^4.0|^5.0",
-    "symfony/yaml": "^2.7|^3.0|^4.0|^5.0",
-    "symfony/config": "^2.7|^3.0|^4.0|^5.0",
-    "symfony/http-kernel": "^2.7|^3.0|^4.0|^5.0",
+    "php": "^7.2.5",
+    "symfony/dependency-injection": "^5.0",
+    "symfony/yaml": "^5.0",
+    "symfony/config": "^5.0",
+    "symfony/http-kernel": "^5.0",
     "twig/twig": "^1.18|^2.0",
     "doctrine/annotations": "^1.0",
     "flagception/flagception": "^1.5"
   },
   "require-dev": {
-    "phing/phing": "^2.15",
-    "symfony/phpunit-bridge": ">=2.7",
-    "symfony/twig-bridge": ">=2.7",
-    "flagception/contentful-activator": "^1.0",
+    "symfony/phpunit-bridge": "^5.0",
+    "symfony/twig-bridge": "^5.0",
     "flagception/database-activator": "^1.0",
-    "phpunit/phpunit": "^5.5",
-    "php-mock/php-mock-phpunit": "^1.1",
-    "squizlabs/php_codesniffer": "^2.8",
+    "phpunit/phpunit": "8.3.*",
+    "php-mock/php-mock-phpunit": "^2.2",
+    "squizlabs/php_codesniffer": "^3.0",
     "php-coveralls/php-coveralls": "^2.0"
   },
   "autoload": {
@@ -50,5 +48,11 @@
   "suggest": {
     "flagception/contentful-activator": "Fetch feature flags from contentful.",
     "flagception/database-activator": "Fetch feature flags from a database."
+  },
+  "scripts": {
+    "tests": [
+      "./vendor/bin/phpcs",
+      "./vendor/bin/phpunit --coverage-text"
+    ]
   }
 }
