@@ -28,7 +28,7 @@ class FlagceptionExtensionTest extends TestCase
     /**
      * {@inheritdoc}
      */
-    protected function setUp()
+    protected function setUp(): void
     {
         $container = new ContainerBuilder();
         $loader = new Loader\YamlFileLoader($container, new FileLocator(__DIR__ . '/../../src/Resources/config'));
